--- a/libs/librepcb/editor/project/boardeditor/fabricationoutputdialog.cpp
+++ b/libs/librepcb/editor/project/boardeditor/fabricationoutputdialog.cpp
@@ -191,7 +191,7 @@ void FabricationOutputDialog::btnGenerateClicked() {
 
     // rebuild planes because they may be outdated!
     BoardPlaneFragmentsBuilder builder;
-    builder.runSynchronously(mBoard);  // can throw
+    builder.runAndApply(mBoard);  // can throw
 
     // update fabrication output settings if modified
     BoardFabricationOutputSettings s = mBoard.getFabricationOutputSettings();