--- a/tests/unittests/core/project/board/boardgerberexporttest.cpp
+++ b/tests/unittests/core/project/board/boardgerberexporttest.cpp
@@ -80,7 +80,7 @@ TEST(BoardGerberExportTest, test) {
 
   // force planes rebuild
   BoardPlaneFragmentsBuilder builder;
-  builder.runSynchronously(*board);  // can throw
+  builder.runAndApply(*board);  // can throw
 
   // export fabrication data
   BoardFabricationOutputSettings config = board->getFabricationOutputSettings();