… libs/librepcb/core/CMakeLists.txt
… libs/librepcb/core/project/board/drc/boardclipperpathgenerator.cpp
… libs/librepcb/core/project/board/drc/boarddesignrulecheck.h
… libs/librepcb/core/project/board/drc/boarddesignrulecheckdata.cpp
… libs/librepcb/core/project/board/drc/boarddesignrulecheckdata.h
… libs/librepcb/core/project/board/drc/boarddesignrulecheckmessages.cpp
… libs/librepcb/core/project/board/drc/boarddesignrulecheckmessages.h
… libs/librepcb/editor/project/boardeditor/boardeditor.cpp
… tests/unittests/CMakeLists.txt
… tests/unittests/core/project/board/boarddesignrulechecktest.cpp
… tests/unittests/core/project/board/boardplanefragmentsbuildertest.cpp