This PR refactors two CPU intensive operations to run them now with multiple threads:
- Plane fragment building
- Design rule check

Planes are now using a separate thread for each layer. So the more layers, the more time is saved now. In a real project with planes on 4 layers, I measured an improvement from ~170ms to ~95ms (44%), and with a complex 2-layer board I measured an improvement from 490ms to 330ms (33%). This makes the UI feeling more responsive while drawing traces etc.

For the DRC, it is more complicated since there are many different kinds of checks and some checks depend on CPU-intensive precalculated data. The concept is documented in the source code:

```
  // - A subset of jobs, called "stage 1", is run in parallel to calculate data
  //   which other jobs depend on (e.g. copper areas on each layer).
  // - Jobs depending on this data, called "stage 2", are started after all
  //   jobs of stage 1 completed. Stage 2 jobs are run in parallel then too.
  // - A third set of jobs, called "independent", does not depend on stage 1
  //   jobs output data and is thus run in parallel to stage 1 & 2 jobs.
  // - Very trivial (CPU inexpensive) jobs are run sequentially in this thread
  //   to avoid spawning a large amount of threads. They are run sequentially,
  //   but in parallel to stage 2 jobs since this thread has no other work to
  //   do then.
  //
  //        ▲                           ┌────────────────────────────────┐
  //        │                         ┌►│        Independent jobs        │
  //        │                         │ └────────────────────────────────┤
  // Threads│                         │                                  │
  //  2..n  │                         │ ┌────────────┐   ┌───────────────┤
  //        │                         ├►│Stage 1 jobs│ ┌►│ Stage 2 jobs  │
  //        │                         │ └────────────┤ │ └───────────────┤
  //        │                         │              ▼ │                 │
  //        │              ┌──────────┤              ┌─┤ ┌───────────────┤
  //  run() │            ┌►│Spawn jobs│--------------│ ├►│Sequential jobs│
  // Thread │            │ └──────────┘              └─┘ └───────────────┤
  //        │            │                                               ▼
  //        │ ┌──────────┤                                               ┌───┐
  //  Main  │ │Copy board│-----------------------------------------------│End│
  // Thread │ └──────────┘                                               └───┘
  //        └────────────────────────────────────────────────────────────────► t
```

For a complex 2-layer board, I measured an improvement from 14.2s to 7.5s (47%), which is very nice since the DRC is an operation the user needs to wait for its result.

In addition to the speedup, the refactoring also fixes a few minor DRC bugs and adds initial unit tests for the DRC.